requirements.txt
colorama==0.4.6
cryptography==38.0.4
Django==4.1.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
numpy==1.23.5
opencv_python==4.6.0.66
pandas==1.5.2
Pillow==9.3.0
pure_python_adb==0.3.0.dev0
PyDrive==1.3.1
pyTelegramBotAPI==4.7.1
pytesseract==0.3.10
PyYAML==6.0
requests==2.28.1
termcolor==2.1.1
Unidecode==1.3.6
websocket_client==1.4.2
